finished new modal, delete old (areyousure
