--- components/AreYouSure/AreYouSure.css
@@ -1,27 +1,0 @@
-.areYouSure {
-  background-color: lemonchiffon;
-  padding: 0.5rem;
-  border: 2px solid var(--red);
-  border-radius: 4px;
-  display: block;
-  position: absolute;
-  top: 300px;
-  left: 5%;
-  height: fit-content;
-  width: 90%;
-  z-index: 2;
-}
-
-.areYouSure--active {
-  display: block;
-}
-
-.areYouSure--passive {
-  display: none;
-}
-
-.no_button {
-  font-weight: 700;
-  border: 2px solid var(--red);
-  background-color: lemonchiffon;
-}
