+++ components/Modal/Modal.css
@@ -15,3 +15,14 @@
 .modal--passive {
   display: none;
 }
+
+.yes_button {
+  color: var(--white);
+  font-weight: 700;
+  border: 2px solid var(--red);
+  background-color: var(--red);
+}
+
+.button--passive {
+  display: none;
+}
